requirements.txt
Django>=5.0,<6.0
psycopg2-binary
python-decouple==3.8.0
django-model-utils
djangorestframework
djangorestframework-simplejwt
